Rename cqc pir cleaned column names file
