+++ tests/test_file_schemas.py
@@ -50,7 +50,7 @@ from utils.column_names.raw_data_files.cqc_pir_columns import (
 from utils.column_names.raw_data_files.ascwds_workplace_columns import (
     AscwdsWorkplaceColumns as AWP,
 )
-from utils.column_names.cleaned_data_files.cqc_pir_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_pir_cleaned import (
     CqcPIRCleanedColumns as CQCPIRClean,
 )
 from utils.column_names.cleaned_data_files.cqc_provider_cleaned_values import (
