+++ utils/column_values/categorical_columns_by_dataset.py
@@ -4,7 +4,7 @@ from utils.column_names.cleaned_data_files.cqc_location_cleaned import (
     CqcLocationCleanedColumns as CQCLClean,
     NewCqcLocationApiColumns as CQCL,
 )
-from utils.column_names.cleaned_data_files.cqc_pir_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_pir_cleaned import (
     CqcPIRCleanedColumns as PIRClean,
 )
 from utils.column_names.cleaned_data_files.cqc_provider_cleaned_values import (
