+++ utils/validation/validation_rules/pir_cleaned_validation_rules.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from utils.column_names.cleaned_data_files.cqc_pir_cleaned_values import (
+from utils.column_names.cleaned_data_files.cqc_pir_cleaned import (
     CqcPIRCleanedColumns as CQCPIRClean,
 )
 from utils.column_values.categorical_columns_by_dataset import (
